Add residual reduction algorithm
Two implementations of residual reductions were added to the reduction
function. There is possible parametrization using the ParameterMap to
select the type of implementation and its direction. The types differ in
 whether the construction of residual automaton is done after the
final determinization or during the determinization. Both implementations
had similar results in effectivity and the result automata are mostly
the same, but the approaches may provide further inspiration and
different optimizations.

However, we were not able to find a way to utilize StateRenaming as it
is done for simulation, as for residual reduction the automaton is
reverted and determinized twice and there is no possible direct
connection between the state of the original and tthe result automaton.
For this reason, the parameter is not used.

Testing of this new functionality is also provided on various-sized
automaton, checking the result construction. In every case, the size of
both approaches should be the same(applies only for the same direction)
and on some automata the results are identical.